artportfolio/requirements.txt
Django==5.2.6
pillow==11.3.0
djangorestframework==3.16.1
requests==2.32.5
django-allauth==65.11.1
dj-rest-auth==7.0.1